requirements.txt
Django
django-ckeditor
Pillow
validate-email
requests
environ
